feat: 댓글 작성자, 비밀번호 검증 기능 추가

+++ src/main/java/com/blog/be/comment/application/CommentCommandService.java
@@ -18,32 +18,40 @@ public class CommentCommandService {
     private final CommentRepository commentRepository;
 
     // 댓글 달기
-    public void createRootComment(Long postId, String content) {
-        CommentJpaEntity comment = CommentJpaEntity.createRoot(postId, content);
+    public void createRootComment(Long postId, String author, String password, String content) {
+        CommentJpaEntity comment = CommentJpaEntity.createRoot(postId, author, password, content);
 
         commentRepository.save(comment);
     }
 
     // 대댓글 작성
-    public void createReply(Long parentId, Long postId, String content) {
+    public void createReply(Long parentId, Long postId, String author, String password, String content) {
         Long parentCommentId = commentRepository.findById(parentId)
                 .map(CommentJpaEntity::getId)
                 .orElseThrow(() -> new CommentException(CommentErrorCode.PARENT_COMMENT_NOT_FOUND));
 
-        CommentJpaEntity newReply = CommentJpaEntity.createReply(postId, parentCommentId, content);
+        CommentJpaEntity newReply = CommentJpaEntity.createReply(postId, parentCommentId, author, password, content);
         commentRepository.save(newReply);
     }
 
     // 댓글 수정
-    public void updateComment(Long id, String newContent) {
+    public void updateComment(Long id, String password, String newContent) {
         CommentJpaEntity comment = getComment(id);
+
+        if (!comment.isPasswordMatch(password)) {
+            throw new CommentException(CommentErrorCode.INVALID_PASSWORD);
+        }
 
         comment.changeContent(newContent);
     }
 
     // 댓글 삭제
-    public void deleteComment(Long id) {
+    public void deleteComment(Long id, String password) {
         CommentJpaEntity comment = getComment(id);
+
+        if (!comment.isPasswordMatch(password)) {
+            throw new CommentException(CommentErrorCode.INVALID_PASSWORD);
+        }
 
         commentRepository.delete(comment);
     }
